python unittests work after refractor; mypy check passes
  Re-enabled mypy in Github actions. Fixes #14
