+++ .github/workflows/all.yml
@@ -91,26 +91,25 @@ jobs:
   #    run: |
   #      black --check asciiserialcom tests
 
-  ## Will fix issues later and re-enable
-  #python-mypy:
-  #  runs-on: ubuntu-latest
-  #  defaults:
-  #    run:
-  #      working-directory: python
-  #  steps:
-  #  - uses: actions/checkout@v2
-  #  - name: Set up Python 3.9
-  #    uses: actions/setup-python@v2
-  #    with:
-  #      python-version: 3.9
-  #  - name: Install dependencies
-  #    run: |
-  #      python -m pip install --upgrade pip
-  #      pip install mypy
-  #      pip install .
-  #  - name: Static Analysis with mypy
-  #    run: |
-  #      mypy asciiserialcom
+  python-mypy:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: python
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.9
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.9
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install mypy
+        pip install .
+    - name: Static Analysis with mypy
+      run: |
+        mypy asciiserialcom tests
 
   python-unittest:
     strategy:
