+++ python/tests/__init__.py
@@ -1,5 +1,7 @@
-from .test_helpers import *
+# type: ignore
+
 from .test_message import *
-from .test_asciiSerialCom import *
+from .test_base import *
+from .test_host import *
 from .test_circularBuffer import *
 from .test_integration import *
